+++ cli/src/plunder.ts
@@ -9,10 +9,17 @@ import {
     InvalidCaptureScreenshotsOption,
 } from '@eighty4/plunder-core'
 
-const ascii = {
+const ansi = {
     bold: (s: string) => `\u001b[1m${s}\u001b[0m`,
     underline: (s: string) => `\u001b[4m${s}\u001b[0m`,
     red: (s: string) => `\u001b[31m${s}\u001b[0m`,
+    rewriteLines: (n: number, s: string) => {
+        let seq = ''
+        for (let i = 0; i < n; i++) {
+            seq += `\u001b[A`
+        }
+        console.log(seq + s)
+    }
 }
 
 const knownOpts = {
@@ -50,17 +57,31 @@ const parsed = nopt(knownOpts, shortHands)
 if (parsed.help) {
     helpPrint()
 } else {
-    const opts = {
+    const opts: CaptureScreenshotsOptions = {
         browser: parsed['browser'] ?? 'chromium',
         devices: parsed['devices'] === false ? false : parsed['device']?.length ? parsed['device'] : true,
         outDir: parsed['out-dir'],
         headless: parsed['not-headless'] !== true,
         recursive: parsed['recursive'] === true,
         urls: parsed.argv.remain,
+
+        progress: function (update) {
+            switch (update.step) {
+                case 'parsing':
+                    ansi.rewriteLines(1, `Parsing CSS: ${update.pages.completed} / ${update.pages.total} pages parsed`)
+                    break
+                case 'capturing':
+                    ansi.rewriteLines(1, `Capturing screenshots: ${update.screenshots.completed} / ${update.screenshots.total} screenshots captured`)
+                    break
+                case 'completed':
+                    ansi.rewriteLines(1, 'Capturing screenshots completed!')
+            }
+        }
     }
 
     await validateOpts(opts)
 
+    console.log('Get ready to plunder!')
     captureScreenshots(opts).then().catch(onError)
 }
 
@@ -102,9 +123,9 @@ function onError(e: any) {
 }
 
 function errorPrint(s: string, optParseError?: boolean) {
-    console.log(ascii.bold(ascii.red('error:')), s)
+    console.log(ansi.bold(ansi.red('error:')), s)
     if (optParseError) {
-        console.log(`\nFor more information, try '${ascii.bold('--help')}'.`)
+        console.log(`\nFor more information, try '${ansi.bold('--help')}'.`)
     }
     process.exit(1)
 }
@@ -112,15 +133,15 @@ function errorPrint(s: string, optParseError?: boolean) {
 function helpPrint() {
     console.log(`Plunders CSS of websites and outputs screenshots around media query breakpoints.
 
-${ascii.bold(ascii.underline('Usage:'))} ${ascii.bold('plunder')} [OPTIONS] URL...
+${ansi.bold(ansi.underline('Usage:'))} ${ansi.bold('plunder')} [OPTIONS] URL...
 
-${ascii.bold(ascii.underline('Options:'))}
-  ${ascii.bold('-b')}, ${ascii.bold('--browser')} <BROWSER>    Browser engine [values: chromium (default) | firefox | webkit]
-  ${ascii.bold('-d')}, ${ascii.bold('--device')} <DEVICE>      Device name patterns to use for screenshot capturing
-      ${ascii.bold('--no-devices')}           Excludes device emulation for screenshot capturing [defaults to false]
-      ${ascii.bold('--not-headless')}         Launch browser as a headless process [defaults to false]
-  ${ascii.bold('-o')}, ${ascii.bold('--out-dir')} <OUT_DIR>    Output directory for screenshot capture [required]
-  ${ascii.bold('-r')}, ${ascii.bold('--recursive')}            Recursively plunders links to pages on same domain [defaults to false]
-  ${ascii.bold('-h')}, ${ascii.bold('--help')}                 Print help`)
+${ansi.bold(ansi.underline('Options:'))}
+  ${ansi.bold('-b')}, ${ansi.bold('--browser')} <BROWSER>    Browser engine [values: chromium (default) | firefox | webkit]
+  ${ansi.bold('-d')}, ${ansi.bold('--device')} <DEVICE>      Device name patterns to use for screenshot capturing
+      ${ansi.bold('--no-devices')}           Excludes device emulation for screenshot capturing [defaults to false]
+      ${ansi.bold('--not-headless')}         Launch browser as a headless process [defaults to false]
+  ${ansi.bold('-o')}, ${ansi.bold('--out-dir')} <OUT_DIR>    Output directory for screenshot capture [required]
+  ${ansi.bold('-r')}, ${ansi.bold('--recursive')}            Recursively plunders links to pages on same domain [defaults to false]
+  ${ansi.bold('-h')}, ${ansi.bold('--help')}                 Print help`)
     process.exit(0)
 }